+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `#[builder(build_method(vis="pub", name=build))]` for customizing visibility and fn name of the final build method
+  (the default visibility is `pub`, and default build name is `build`)
 
 ## 0.10.0 - 2022-02-13
 ### Added